Add files via upload
This programs were added or modified so that now the program can produce more wig files that will have the amount of N's in the kmers and another track were only mismatches are going to be shown (mismatches are were there are different bases and not were there are N's). So now the program can help to differentiate if a region of low conservation is because of mismatches or just because the sequencing of that part is not completed.